added RLBotConfig, and reworked access modifiers and method naming for both CFE and RLBotConfig
