+++ test/dk/aau/cs/ds306e18/tournament/utility/configuration/ConfigFileEditorTest.java
@@ -1,22 +1,16 @@
 package dk.aau.cs.ds306e18.tournament.utility.configuration;
 
-import dk.aau.cs.ds306e18.tournament.model.Bot;
-import dk.aau.cs.ds306e18.tournament.model.Team;
-import dk.aau.cs.ds306e18.tournament.model.match.Match;
-import dk.aau.cs.ds306e18.tournament.utility.configuration.ConfigFileEditor;
 import org.junit.Test;
 
 import java.util.ArrayList;
 
-import static junit.framework.TestCase.assertEquals;
-import static junit.framework.TestCase.assertNull;
-import static junit.framework.TestCase.assertTrue;
+import static junit.framework.TestCase.*;
 import static org.junit.Assert.assertFalse;
 
 public class ConfigFileEditorTest {
 
-    private final static String testDir = "test/";
-    private final static String testConfigFilename = "testconfig.cfg";
+    final static String testDir = "test/";
+    final static String testConfigFilename = "testconfig.cfg";
     private ArrayList<String> testConfig = new ArrayList<>();
 
     /**
@@ -28,44 +22,13 @@ public class ConfigFileEditorTest {
         testConfig.add("participant_team_1 = 1");
     }
 
-    /**
-     * Creates a simple Match with three bots on each team
-     *
-     * @return created match
-     */
-    private static Match createTestMatch() {
-        Bot blueBot1 = new Bot("Blue1", "BlueDev1", "path/blue/1");
-        Bot blueBot2 = new Bot("Blue2", "BlueDev2", "path/blue/2");
-        Bot blueBot3 = new Bot("Blue3", "BlueDev3", "path/blue/3");
-
-        ArrayList<Bot> blueBots = new ArrayList<>();
-        blueBots.add(blueBot1);
-        blueBots.add(blueBot2);
-        blueBots.add(blueBot3);
-
-        Team blueTeam = new Team("BlueTeam", blueBots, 0, "BlueDescription");
-
-        Bot orangeBot1 = new Bot("Orange1", "OrangeDev1", "path/orange/1");
-        Bot orangeBot2 = new Bot("Orange2", "OrangeDev2", "path/orange/2");
-        Bot orangeBot3 = new Bot("Orange3", "OrangeDev3", "path/orange/3");
-
-        ArrayList<Bot> orangeBots = new ArrayList<>();
-        orangeBots.add(orangeBot1);
-        orangeBots.add(orangeBot2);
-        orangeBots.add(orangeBot3);
-
-        Team orangeTeam = new Team("OrangeTeam", orangeBots, 1, "OrangeDescription");
-
-        return new Match(blueTeam, orangeTeam);
-    }
-
     @Test
     public void readWriteConcurrencyTest() {
         setTestConfig();
         ConfigFileEditor.setConfig(testConfig);
-        ConfigFileEditor.writeConfig(testDir + "dummyconfig.cfg");
+        ConfigFileEditor.write(testDir + "dummyconfig.cfg");
 
-        ConfigFileEditor.readConfig(testDir + "dummyconfig.cfg");
+        ConfigFileEditor.read(testDir + "dummyconfig.cfg");
         assertEquals(testConfig, ConfigFileEditor.getConfig());
     }
 
@@ -135,48 +98,14 @@ public class ConfigFileEditorTest {
     }
 
     @Test
-    public void configureMatchTest1() {
-        ConfigFileEditor.readConfig(testDir + testConfigFilename);
-        Match match = createTestMatch();
-        ConfigFileEditor.configureMatch(match);
-
-
-        assertEquals("6", ConfigFileEditor.getValueOfLine("num_participant"));
-
-        // checking correctly set paths
-        assertEquals(match.getBlueTeam().getBots().get(0).getConfigPath(),
-                ConfigFileEditor.getValueOfLine("participant_config_0"));
-        assertEquals(match.getBlueTeam().getBots().get(1).getConfigPath(),
-                ConfigFileEditor.getValueOfLine("participant_config_1"));
-        assertEquals(match.getBlueTeam().getBots().get(2).getConfigPath(),
-                ConfigFileEditor.getValueOfLine("participant_config_2"));
-
-        assertEquals(match.getOrangeTeam().getBots().get(0).getConfigPath(),
-                ConfigFileEditor.getValueOfLine("participant_config_3"));
-        assertEquals(match.getOrangeTeam().getBots().get(1).getConfigPath(),
-                ConfigFileEditor.getValueOfLine("participant_config_4"));
-        assertEquals(match.getOrangeTeam().getBots().get(2).getConfigPath(),
-                ConfigFileEditor.getValueOfLine("participant_config_5"));
-
-        // checking correctly set team positions
-        for (int i = 0; i < match.getBlueTeam().size(); i++) {
-            assertEquals("0", ConfigFileEditor.getValueOfLine("participant_team_" + i));
-        }
-
-        for (int i = match.getBlueTeam().size(); i < match.getOrangeTeam().size(); i++) {
-            assertEquals("1", ConfigFileEditor.getValueOfLine("participant_team_" + i));
-        }
-    }
-
-    @Test
     public void validateConfigSyntax1() {
-        ConfigFileEditor.readConfig(testDir + testConfigFilename);
+        ConfigFileEditor.read(testDir + testConfigFilename);
         assertTrue(ConfigFileEditor.isValid());
     }
 
     @Test
     public void validateConfigSyntax2() {
-        ConfigFileEditor.readConfig(testDir + testConfigFilename);
+        ConfigFileEditor.read(testDir + testConfigFilename);
         ArrayList<String> config = ConfigFileEditor.getConfig();
         // remove trailing, closing square bracket from first header
         config.set(0, config.get(0).split("]")[0]);
